requirements.txt
openai==1.35.3
annotated-types
charset-normalizer
et-xmlfile
flask-cors
fonttools
pdfminer.six
pillow
pydantic-core
pymupdf
pymupdfb
pypdf2
pysimplegui
pysocks
python-dateutil
scikit-learn
trio-websocket
types-pytz
typing-extensions
praw
streamlit-authenticator
nltk
pyyaml